pyproject.toml
[tool.poetry]
name = "bitswired-api"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{ include = "app", from = "src" }]

[tool.poetry.dependencies]
python = "^3.11"

# [tool.poetry.scripts]
fastapi = { extras = ["all"], version = "^0.109.0" }
sqlalchemy = { extras = ["asyncio"], version = "^2.0.25" }
python-dotenv = "^1.0.1"
pydantic-settings = "^2.1.0"
httpx = "^0.26.0"
jinja2 = "^3.1.3"
aiosqlite = "^0.20.0"
bcrypt = "^4.1.2"
litellm = "^1.34.38"
sse-starlette = "^2.1.0"
markdown = "^3.6"


[tool.poetry.group.dev.dependencies]
alembic = "^1.13.1"
ruff = "^0.1.7"
pytest = "^7.4.3"
pre-commit = "^3.6.0"
commitizen = "^3.13.0"
poethepoet = "^0.25.0"
pytailwindcss = "^0.2.0"

[tool.poe.tasks]
pre-com-install-base = "poetry run pre-commit install"
pre-com-msg-install = "poetry run pre-commit install --hook-type commit-msg"
pre-com-install = ["pre-com-install-base", "pre-com-msg-install"]
db-push.script = "app.db:init_models"
dev = "poetry run uvicorn app.app:app --host 0.0.0.0 --reload"
dev-tailwind = "poetry run tailwindcss -i static/input.css -o static/output.css --watch=always"


[tool.commitizen]
name = "cz_conventional_commits"
tag_format = "$version"
version_scheme = "semver"
version_provider = "poetry"
update_changelog_on_bump = true
major_version_zero = true
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.ruff]
# Same as Black.
line-length = 88
indent-width = 4
# Assume Python 3.11
target-version = "py311"
src = ["src", "tests"]

[tool.ruff.lint]
extend-select = ["I"]

[tool.ruff.format]
# Like Black, use double quotes for strings.
quote-style = "double"
# Like Black, indent with spaces, rather than tabs.
indent-style = "space"
# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false
# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"

[tool.mypy]
exclude = ['src/']
warn_return_any = true
warn_unused_configs = true
disallow_incomplete_defs = true
disallow_untyped_defs = true
ignore_missing_imports = true

[tool.dagster]
module_name = "dags"
